requirements.txt
# Pixer Web Interface Dependencies

# Web Framework
Flask>=3.0.0
Werkzeug>=3.0.0

# Image Processing
Pillow>=10.2.0

# HTTP and Networking (included with Python, but listed for clarity)
# socket - built-in
# threading - built-in
# time - built-in
# logging - built-in

# Optional: For production deployment
# gunicorn>=21.2.0
# waitress>=2.1.2
